+++ im_model/lib/src/equality/equality_hash.dart
@@ -1,54 +1,34 @@
-import 'package:collection/collection.dart';
-
-const _dce = DeepCollectionEquality();
-
 /// Jenkins Hash function (1 to 1)
 /// https://en.wikipedia.org/wiki/Jenkins_hash_function
 class Hash {
   const Hash();
 
-  static int _add(int newHash, Object? o) {
-    var hash = (newHash + o.hashCode) & 0x7fffffff;
-    hash = (hash + (hash << 10)) & 0x7fffffff;
+  @pragma("vm:prefer-inline")
+  int hash(Object obj, List<Object?> props) {
+    final hash = _finish(props.fold(0, (h, i) => _combine(h, i.hashCode)));
+
+    return obj.runtimeType.hashCode ^ hash;
+  }
+
+  /// Generates a hash code for multiple [objects] in iterable.
+  int hashIterable(Iterable<dynamic> it) {
+    return _finish(it.fold(0, (h, i) => _combine(h, i.hashCode)));
+  }
+
+  /// Generates a hash code for two objects.
+  int hash2(dynamic a, dynamic b) {
+    return _finish(_combine(_combine(0, a.hashCode), b.hashCode));
+  }
+
+  int _combine(int hash, int value) {
+    hash = 0x1fffffff & (hash + value);
+    hash = 0x1fffffff & (hash + ((0x0007ffff & hash) << 10));
     return hash ^ (hash >> 6);
   }
 
-  static int _finish(int newHash) {
-    var hash = (newHash + (newHash << 3)) & 0x7fffffff;
-    hash ^= hash >> 11;
-    return (hash + (hash << 15)) & 0x7fffffff;
-  }
-
-  @pragma("vm:prefer-inline")
-  int hash(Object obj, List<Object?> props) {
-    var hash = 0;
-    for (final hashParam in props) {
-      if (hashParam is Set) {
-        hash = _add(
-          hash,
-          const SetEquality(_dce).hash(hashParam) ^ hashParam.length,
-        );
-      } else if (hashParam is Map) {
-        hash = _add(
-          hash,
-          const MapEquality(keys: _dce, values: _dce).hash(hashParam) ^
-              hashParam.length,
-        );
-      } else if (hashParam is List) {
-        hash = _add(
-          hash,
-          const ListEquality(_dce).hash(hashParam) ^ hashParam.length,
-        );
-      } else if (hashParam is Iterable) {
-        hash = _add(
-          hash,
-          const IterableEquality(_dce).hash(hashParam) ^ hashParam.length,
-        );
-      } else {
-        hash = _add(hash, _dce.hash(hashParam));
-      }
-    }
-
-    return obj.runtimeType.hashCode ^ _finish(hash);
+  int _finish(int hash) {
+    hash = 0x1fffffff & (hash + ((0x03ffffff & hash) << 3));
+    hash = hash ^ (hash >> 11);
+    return 0x1fffffff & (hash + ((0x00003fff & hash) << 15));
   }
 }